% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{assign.global}
\alias{assign.global}
\title{function loading ncores in global environment}
\usage{
assign.global(ncores, pos = 1)
}
\arguments{
\item{ncores}{ncores to assign to global}

\item{pos}{defaults to 1 which equals an assignment to global environment}
}
\description{
function loading ncores in global environment
}
\keyword{internal}
