% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.R
\name{explore}
\alias{explore}
\title{Function to explore bootstrapped results of a SEQuential object}
\usage{
explore(obj, n)
}
\arguments{
\item{obj}{SEQuential object to explore}

\item{n}{bootstrap iteration to return}
}
\value{
a SEQuential object where outcome_model and weight_statistics are limited to the bootstrap \code{n}
}
\description{
Function to explore bootstrapped results of a SEQuential object
}
