% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fatglmHelpers.R
\name{fastglm.clean}
\alias{fastglm.clean}
\title{Function to clean out non needed elements from fastglm return}
\usage{
\method{fastglm}{clean}(model)
}
\arguments{
\item{model}{a fastglm model}
}
\description{
Function to clean out non needed elements from fastglm return
}
\keyword{internal}
