% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fatglmHelpers.R
\name{fastglm.robust}
\alias{fastglm.robust}
\title{Function to calculate robust standard errors from a fastglm model}
\usage{
\method{fastglm}{robust}(model, X, y, weight = NULL)
}
\description{
Function to calculate robust standard errors from a fastglm model
}
\keyword{internal}
