% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{inline.pred}
\alias{inline.pred}
\title{Helper Function to inline predict a fastglm object}
\usage{
inline.pred(model, newdata, params, type, case = "default")
}
\arguments{
\item{model}{a fastglm object}

\item{newdata}{filler for a .SD from data.table}

\item{params}{parameter from SEQuential}

\item{type}{type of prediction}
}
\description{
Helper Function to inline predict a fastglm object
}
\keyword{internal}
