% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_test.R
\name{multinomial.predict}
\alias{multinomial.predict}
\title{Helper to predict from the nested logistic}
\usage{
multinomial.predict(model, X)
}
\description{
Helper to predict from the nested logistic
}
\keyword{internal}
