% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{prediction.passer}
\alias{prediction.passer}
\title{Passes model to correct prediction method}
\usage{
prediction.passer(model, X, params, type)
}
\description{
Passes model to correct prediction method
}
\keyword{internal}
