% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_setters.R
\name{prepare.output}
\alias{prepare.output}
\title{Output constructor}
\usage{
prepare.output(
  params,
  outcome_model,
  hazard,
  robustSE,
  survival_plot,
  survival_data,
  risk,
  elapsed_time,
  info
)
}
\description{
Output constructor
}
\keyword{internal}
