% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.R
\name{risk.ratio}
\alias{risk.ratio}
\title{Function to return risk ratios from a SEQuential object}
\usage{
risk.ratio(object)

risk.ratio(object)
}
\arguments{
\item{object}{SEQoutput object}
}
\value{
list of risk ratios

list of risk differences
}
\description{
Function to return risk ratios from a SEQuential object

Function to return risk differences from a SEQuential object
}
