% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.R
\name{risk}
\alias{risk}
\title{Function to return risk information from a SEQuential object}
\usage{
risk(object)
}
\arguments{
\item{object}{SEQoutput object}
}
\value{
a dataframe of Risk information (risk ratios, risk differences and confidence intervals, if bootstrapped)
}
\description{
Function to return risk information from a SEQuential object
}
