% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.R
\name{robust}
\alias{robust}
\title{Function to return robust standard errors from a SEQuential object}
\usage{
robust(object)
}
\arguments{
\item{object}{SEQoutput object}
}
\value{
list of robust standard errors of outcome models
}
\description{
Function to return robust standard errors from a SEQuential object
}
