% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fatglmHelpers.R
\name{robust_fastglm}
\alias{robust_fastglm}
\title{Function to calculate robust standard errors from a fastglm model}
\usage{
robust_fastglm(model, X, y, weight = NULL)
}
\description{
Function to calculate robust standard errors from a fastglm model
}
\keyword{internal}
