% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEQopts.R
\name{SEQopts}
\alias{SEQopts}
\title{Parameter Builder - passing defaults to \code{SEQuential}}
\usage{
SEQopts(
  parallel = FALSE,
  nthreads = data.table::getDTthreads(),
  ncores = parallel::detectCores() - 1,
  bootstrap = FALSE,
  nboot = 100,
  boot.sample = 0.8,
  seed = 1636,
  min.followup = -Inf,
  max.followup = Inf,
  max.survival = Inf,
  include.period = TRUE,
  include.trial = TRUE,
  covariates = NA,
  weighted = FALSE,
  upper.weight = Inf,
  lower.weight = -Inf,
  p99.weight = FALSE,
  numerator = NA,
  denominator = NA,
  surv = NA,
  pre.expansion = TRUE,
  elig.wts.0 = NA,
  elig.wts.1 = NA,
  hazard = FALSE,
  calculate.var = FALSE,
  random.selection = FALSE,
  selection.prob = 0.8,
  ltfu.numerator = NA,
  ltfu.denominator = NA,
  cense = NA,
  cense2 = NA,
  eligible_cense = NA,
  eligible_cense2 = NA,
  compevent = NA,
  multinomial = FALSE,
  treat.level = c(0, 1),
  excused = FALSE,
  excused.col1 = NA,
  excused.col0 = NA,
  km.curves = FALSE,
  baseline.indicator = "_bas",
  squared.indicator = "_sq",
  followup.class = FALSE,
  followup.spline = 0,
  fastglm.method = 2L
)
}
\arguments{
\item{parallel}{Logical: define if the SEQuential process is run in parallel, default is FALSE}

\item{nthreads}{Integer: number of threads to use for data.table processing}

\item{ncores}{Integer: number of cores to use in parallel processing, default is one less than system max}

\item{bootstrap}{Logical: defines if SEQuential should run bootstrapping, default is FALSE}

\item{nboot}{Integer: number of bootstraps}

\item{boot.sample}{Numeric: percentage of data to use when bootstrapping, should in [0, 1], default is 0.8}

\item{seed}{Integer: starting seed}

\item{min.followup}{Numeric: minimum time to expand aboud, default is -Inf (no minimum)}

\item{max.followup}{Numeric: maximum time to expand about, default is Inf (no maximum)}

\item{max.survival}{Numeric: maximum time for survival curves, default is Inf (no maximum)}

\item{include.period}{Logical: whether or not to include 'period' and 'period_squared' in the outcome model}

\item{include.trial}{Logical: whether or not to include 'trial' and 'trial_squared' in the outcome model}

\item{covariates}{String: covariates to coerce into a formula object, eg. "A+B*C"}

\item{weighted}{Logical: whether or not to preform weighted analysis, default is FALSE}

\item{upper.weight}{Numeric: weights truncated at upper end at this weight}

\item{lower.weight}{Numeric: weights truncated at lower end at this weight}

\item{p99.weight}{Logical: forces weight truncation at 1st and 99th percentile weights, will override provided \code{upper.weight} and \code{lower.weight}}

\item{numerator}{String: numerator covariates to coerce to formula object}

\item{denominator}{String: denominator covariates to coerce to formula object}

\item{surv}{String: survival covariates to coerce to formula object}

\item{pre.expansion}{Logical: whether weighting should be done on pre-expanded data}

\item{elig.wts.0}{String: TODO}

\item{elig.wts.1}{String: TODO}

\item{hazard}{Logical: TODO}

\item{calculate.var}{Logical: TODO}

\item{random.selection}{Logical: TODO}

\item{selection.prob}{Numeric: TODO}

\item{ltfu.numerator}{String: TODO}

\item{ltfu.denominator}{String: TODO}

\item{cense}{String: TODO}

\item{cense2}{String: TODO}

\item{eligible_cense}{String: TODO}

\item{eligible_cense2}{String: TODO}

\item{compevent}{String: TODO}

\item{multinomial}{Logical: whether or not to expect multinomial models}

\item{treat.level}{List: which treatment levels to compare through survival curves}

\item{excused}{Logical: in the case of censoring, whether there is an excused condition}

\item{excused.col1}{String: in the case of \code{excused = TRUE} the column name for Excused1}

\item{excused.col0}{String: in the case of \code{excused = TRUE} the column name for Excused0}

\item{km.curves}{Logical: Kaplan-Meier survival curve creation and data return}

\item{baseline.indicator}{String: identifier for baseline variables in \code{covariates, numerator, denominator} - intended as an override}

\item{squared.indicator}{String: identifier for squared variables in \code{covariates, numerator, denominator} - intended as an override}

\item{followup.class}{Logical: whether or not to treat followup as a class in models}

\item{followup.spline}{Integer: degrees of freedom for natural (restricted) splines for use in trial-time}

\item{fastglm.method}{Integer: decomposition method for fastglm (1-QR, 2-Cholesky, 3-LDLT, 4-QR.FPIV)}
}
\value{
An object of class 'SEQOpts'
}
\description{
Parameter Builder - passing defaults to \code{SEQuential}
}
