% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEQoutput.R
\docType{class}
\name{SEQoutput-class}
\alias{SEQoutput-class}
\title{An S4 class used to hold the outputs for the SEQuential process}
\description{
An S4 class used to hold the outputs for the SEQuential process
}
\section{Slots}{

\describe{
\item{\code{boostrap}}{if the process was bootstrapped}

\item{\code{bootstrap.sample}}{how much of the data was sampled during bootstrapping}

\item{\code{boot.slice}}{when using method 'show' defines what to print}

\item{\code{seed}}{seeding any random sampling occuring in bootstrap process}

\item{\code{bootstrap.nboot}}{number of bootstraps}

\item{\code{outcome}}{outcome covariates}

\item{\code{numerator}}{numerator covariates}

\item{\code{denominator}}{denominator covariates}

\item{\code{outcome_model}}{list of length \code{bootstrap.nboot} containing outcome coefficients}

\item{\code{hazard}}{hazard ratio}

\item{\code{robust_se}}{robust standard errors}

\item{\code{survival_curve}}{ggplot object for the survival curves}

\item{\code{survival_data}}{data.table of survival data}

\item{\code{risk_difference}}{risk difference calculated from survival data}

\item{\code{risk_ratio}}{risk ratio calculated from survival data}

\item{\code{elapsed_time}}{time in minutes used for the SEQuential process}

\item{\code{weight_statistics}}{information from the weighting process, containing weight coefficients and weight statistics}
}}

