% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.R
\docType{class}
\name{SEQweights-class}
\alias{SEQweights-class}
\title{An internal S4 class to help transfer weight statistics out of \code{internal_weights}}
\description{
An internal S4 class to help transfer weight statistics out of \code{internal_weights}
}
\section{Slots}{

\describe{
\item{\code{weights}}{a data.table containing the estimated weights, either pre or post expansion}

\item{\code{coef.n0}}{coefficients from the numerator zero model}

\item{\code{coef.n1}}{coefficients from the numerator one model}

\item{\code{coef.d0}}{coefficients from the denominator zero model}

\item{\code{coef.d1}}{coefficients from the denominator one model}
}}

\keyword{internal}
