% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.R
\name{show,SEQoutput-method}
\alias{show,SEQoutput-method}
\title{TODO}
\usage{
\S4method{show}{SEQoutput}(object)
}
\arguments{
\item{object}{A SEQoutput object - usually generated from \code{SEQuential}}
}
\description{
TODO
}
