% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.R, R/internal_hazard.R
\name{hazard}
\alias{hazard}
\title{Function to return hazard ratios from a SEQuential object}
\usage{
hazard(data, params)

hazard(data, params)
}
\arguments{
\item{data}{expanded data}

\item{params}{SEQuential params}

\item{object}{SEQoutput object}
}
\value{
list of hazard ratios
}
\description{
Function to return hazard ratios from a SEQuential object

Generic function to format a dataset for hazard ratio calculation
}
\keyword{internal}
