% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.R
\name{hazard}
\alias{hazard}
\title{Function to return hazard ratios from a SEQuential object}
\usage{
hazard(object)
}
\arguments{
\item{object}{SEQoutput object}
}
\value{
list of hazard ratios
}
\description{
Function to return hazard ratios from a SEQuential object
}
