% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.R
\docType{class}
\name{SEQopts-class}
\alias{SEQopts-class}
\title{An S4 class of user options to feed into the SEQuential processes and estimates}
\description{
An S4 class of user options to feed into the SEQuential processes and estimates
}
\section{Slots}{

\describe{
\item{\code{parallel}}{Logical: define if the SEQuential process is run in parallel, default is FALSE}

\item{\code{nthreads}}{Integer: number of threads to use for data.table processing}

\item{\code{ncores}}{Integer: number of cores to use in parallel processing, default is one less than system max}

\item{\code{bootstrap}}{Logical: defines if SEQuential should run bootstrapping, default is FALSE}

\item{\code{bootstrap.nboot}}{Integer: number of bootstraps}

\item{\code{bootstrap.sample}}{Numeric: percentage of data to use when bootstrapping, should in [0, 1], default is 0.8}

\item{\code{seed}}{Integer: starting seed}

\item{\code{followup.min}}{Numeric: minimum time to expand aboud, default is -Inf (no minimum)}

\item{\code{followup.max}}{Numeric: maximum time to expand about, default is Inf (no maximum)}

\item{\code{survival.max}}{Numeric: maximum time for survival curves, default is Inf (no maximum)}

\item{\code{followup.include}}{Logical: whether or not to include 'period' and 'period_squared' in the outcome model}

\item{\code{trial.include}}{Logical: whether or not to include 'trial' and 'trial_squared' in the outcome model}

\item{\code{covariates}}{String: covariates to coerce into a formula object, eg. "A+B*C"}

\item{\code{numerator}}{String: numerator covariates to coerce to formula object}

\item{\code{denominator}}{String: denominator covariates to coerce to formula object}

\item{\code{cense.numerator}}{String: TODO}

\item{\code{cense.denominator}}{String: TODO}

\item{\code{surv}}{String: survival covariates to coerce to formula object}

\item{\code{weighted}}{Logical: whether or not to preform weighted analysis, default is FALSE}

\item{\code{weight.lower}}{Numeric: weights truncated at lower end at this weight}

\item{\code{weight.upper}}{Numeric: weights truncated at upper end at this weight}

\item{\code{weight.p99}}{Logical: forces weight truncation at 1st and 99th percentile weights, will override provided \code{weight.upper} and \code{weight.lower}}

\item{\code{weight.eligible0}}{String: TODO}

\item{\code{weight.eligible1}}{String: TODO}

\item{\code{weight.preexpansion}}{Logical: whether weighting should be done on pre-expanded data}

\item{\code{calculate.var}}{Logical: TODO}

\item{\code{hazard}}{Logical: TODO}

\item{\code{selection.random}}{Logical: TODO}

\item{\code{selection.prob}}{Numeric: TODO}

\item{\code{excused}}{Logical: in the case of censoring, whether there is an excused condition}

\item{\code{cense}}{String: TODO}

\item{\code{cense.eligible}}{String: TODO}

\item{\code{multinomial}}{Logical: whether or not to expect multinomial models}

\item{\code{treat.level}}{List: which treatment levels to compare through survival curves}

\item{\code{compevent}}{String: TODO}

\item{\code{excused.col1}}{String: in the case of \code{excused = TRUE} the column name for Excused1}

\item{\code{excused.col0}}{String: in the case of \code{excused = TRUE} the column name for Excused0}

\item{\code{km.curves}}{Logical: Kaplan-Meier survival curve creation and data return}

\item{\code{indicator.baseline}}{String: identifier for baseline variables in \code{covariates, numerator, denominator} - intended as an override}

\item{\code{indicator.squared}}{String: identifier for squared variables in \code{covariates, numerator, denominator} - intended as an override}

\item{\code{fastglm.method}}{Integer: decomposition method for fastglm (1-QR, 2-Cholesky, 3-LDLT, 4-QR.FPIV)}

\item{\code{followup.class}}{Logical: TODO}

\item{\code{followup.spline}}{Logical: TODO}
}}

\keyword{internal}
