% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.R
\docType{class}
\name{SEQopts-class}
\alias{SEQopts-class}
\title{An S4 class of user options to feed into the SEQuential process}
\description{
An S4 class of user options to feed into the SEQuential process
}
\section{Slots}{

\describe{
\item{\code{parallel}}{Logical: to run in parallel}

\item{\code{nthreads}}{Numeric: number of data.table threads to use}

\item{\code{ncores}}{Integer: number of cores to use if parallelized}

\item{\code{bootstrap}}{Logical: defines if SEQuential should run bootstrapping, default is FALSE}

\item{\code{nboot}}{Integer: number of bootstraps}

\item{\code{boot.sample}}{Numeric: percentage of data to use when bootstrapping, should in [0, 1], default is 0.8}

\item{\code{seed}}{Integer: starting seed}

\item{\code{max.followup}}{Numeric: maximum time to expand about, default is Inf (no maximum)}

\item{\code{max.survival}}{Numeric: maximum time for survival curves, default is Inf (no maximum)}

\item{\code{include.period}}{Logical: whether or not to include 'period' and 'period_squared' in the outcome model}

\item{\code{include.trial}}{Logical: whether or not to include 'trial' and 'trial_squared' in the outcome model}

\item{\code{covariates}}{String: covariates to coerce into a formula object, eg. "A+B*C"}

\item{\code{numerator}}{String: numerator covariates to coerce to formula object}

\item{\code{denominator}}{String: denominator covariates to coerce to formula object}

\item{\code{ltfu.numerator}}{String: TODO}

\item{\code{ltfu.denominator}}{String: TODO}

\item{\code{surv}}{String: survival covariates to coerce to formula object}

\item{\code{weighted}}{Logical: whether or not to preform weighted analysis, default is FALSE}

\item{\code{pre.expansion}}{Logical: whether weighting should be done on pre-expanded data}

\item{\code{excused}}{Logical: in the case of censoring, whether there is an excused condition}

\item{\code{cense}}{String: TODO}

\item{\code{cense2}}{String: TODO}

\item{\code{eligible_cense}}{String: TODO}

\item{\code{eligible_cense2}}{String: TODO}

\item{\code{excused.col1}}{String: in the case of \code{excused = TRUE} the column name for Excused1}

\item{\code{excused.col0}}{String: in the case of \code{excused = TRUE} the column name for Excused0}

\item{\code{LTFU}}{Logical: internal flag for LTFU weighting}

\item{\code{baseline.indicator}}{String: identifier for baseline variables in \code{covariates, numerator, denominator} - intended as an override}

\item{\code{squared.indicator}}{String: identifier for squared variables in \code{covariates, numerator, denominator} - intended as an override}

\item{\code{fasgtlm.method}}{Integer: decomposition method for fastglm (1-QR, 2-Cholesky, 3-LDLT, 4-QR.FPIV)}

\item{\code{treat.level}}{List: List of treatments to compare for survival curves}

\item{\code{multinomial}}{Logical: whether or not to preform multinomial analysis}
}}

