% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEQopts.R
\name{SEQopts}
\alias{SEQopts}
\title{Parameter Builder for SEQuential Model and Estimates}
\usage{
SEQopts(
  parallel = FALSE,
  nthreads = data.table::getDTthreads(),
  ncores = parallel::detectCores() - 1,
  bootstrap = FALSE,
  bootstrap.nboot = 100,
  bootstrap.sample = 0.8,
  seed = 1636,
  followup.min = -Inf,
  followup.max = Inf,
  survival.max = Inf,
  followup.include = TRUE,
  trial.include = TRUE,
  covariates = NA,
  weighted = FALSE,
  weight.upper = Inf,
  weight.lower = -Inf,
  weight.p99 = FALSE,
  numerator = NA,
  denominator = NA,
  weight.preexpansion = TRUE,
  weight.eligible0 = NA,
  weight.eligible1 = NA,
  hazard = FALSE,
  calculate.var = FALSE,
  selection.random = FALSE,
  selection.prob = 0.8,
  followup.class = FALSE,
  followup.spline = FALSE,
  cense.numerator = NA,
  cense.denominator = NA,
  cense = NA,
  cense.eligible = NA,
  compevent = NA,
  multinomial = FALSE,
  treat.level = c(0, 1),
  excused = FALSE,
  excused.col1 = NA,
  excused.col0 = NA,
  km.curves = FALSE,
  indicator.baseline = "_bas",
  indicator.squared = "_sq",
  fastglm.method = 2L,
  plot.title = NA,
  plot.subtitle = NA,
  plot.labels = NA,
  plot.colors = c("#F8766D", "#00BFC4"),
  plot.type = "survival"
)
}
\arguments{
\item{parallel}{Logical: define if the SEQuential process is run in parallel, default is FALSE}

\item{nthreads}{Integer: number of threads to use for data.table processing}

\item{ncores}{Integer: number of cores to use in parallel processing, default is one less than system max}

\item{bootstrap}{Logical: defines if SEQuential should run bootstrapping, default is FALSE}

\item{bootstrap.nboot}{Integer: number of bootstraps}

\item{bootstrap.sample}{Numeric: percentage of data to use when bootstrapping, should in [0, 1], default is 0.8}

\item{seed}{Integer: starting seed}

\item{followup.min}{Numeric: minimum time to expand aboud, default is -Inf (no minimum)}

\item{followup.max}{Numeric: maximum time to expand about, default is Inf (no maximum)}

\item{survival.max}{Numeric: maximum time for survival curves, default is Inf (no maximum)}

\item{followup.include}{Logical: whether or not to include 'followup' and 'followup_squared' in the outcome model}

\item{trial.include}{Logical: whether or not to include 'trial' and 'trial_squared' in the outcome model}

\item{covariates}{String: covariates to the right hand side of a formula object}

\item{weighted}{Logical: whether or not to preform weighted analysis, default is FALSE}

\item{weight.upper}{Numeric: weights truncated at upper end at this weight}

\item{weight.lower}{Numeric: weights truncated at lower end at this weight}

\item{weight.p99}{Logical: forces weight truncation at 1st and 99th percentile weights, will override provided \code{weight.upper} and \code{weight.lower}}

\item{numerator}{String: numerator covariates to the right hand side of a  to formula object}

\item{denominator}{String: denominator covariates to the right hand side of a  to formula object}

\item{weight.preexpansion}{Logical: whether weighting should be done on pre-expanded data}

\item{weight.eligible0}{String: column name for indicator column defining which weights are eligible for the 'zero' model}

\item{weight.eligible1}{String: column name for indicator column defining which weights are eligible for the 'one' model}

\item{hazard}{Logical: hazard error calculation instead of survival estimation}

\item{calculate.var}{Logical: robust standard error calculation instead of survival bootstrapping}

\item{selection.random}{Logical: randomly selects IDs with replacement to run analysis}

\item{selection.prob}{Numeric: percent of total IDs to select for \code{selection.random}, should be bound [0, 1]}

\item{followup.class}{Logical: treat followup as a class, e.g. expands every time to it's own indicator column}

\item{followup.spline}{Logical: treat followup as a cubic spline}

\item{cense.numerator}{String: censoring numerator covariates to the right hand side of a formula object}

\item{cense.denominator}{String: censoring denominator covariates to the right hand side of a formula object}

\item{cense}{String: column name for additional censoring variable, e.g. loss-to-follow-up}

\item{cense.eligible}{String: column name for indicator column defining which rows to use for censoring model}

\item{compevent}{String: column name for competing event indicator}

\item{multinomial}{NOT CURRENTLY IMPLEMENTED - WIP}

\item{treat.level}{NOT CURRENTLY IMPLEMENTED - WIP}

\item{excused}{Logical: in the case of censoring, whether there is an excused condition}

\item{excused.col1}{String: in the case of \code{excused = TRUE} the column name for Excused1}

\item{excused.col0}{String: in the case of \code{excused = TRUE} the column name for Excused0}

\item{km.curves}{Logical: Kaplan-Meier survival curve creation and data return}

\item{indicator.baseline}{String: identifier for baseline variables in \code{covariates, numerator, denominator} - intended as an override}

\item{indicator.squared}{String: identifier for squared variables in \code{covariates, numerator, denominator} - intended as an override}

\item{fastglm.method}{Integer: decomposition method for fastglm (1-QR, 2-Cholesky, 3-LDLT, 4-QR.FPIV)}

\item{plot.title}{Character: Title for output plot if \code{km.curves = TRUE}}

\item{plot.subtitle}{Character: Subtitle for output plot if \code{km.curves = TRUE}}

\item{plot.labels}{Character: Color labels for output plot if \code{km.curves = TRUE} in order e.g. \code{c("risk.0", "risk.1")}}

\item{plot.colors}{Character: Colors for output plot if \code{km.curves = TRUE}, defaulted to ggplot2 defaults}

\item{plot.type}{Character: Type of plot to create if \code{km.curves = TRUE}, available options are 'survival', 'risk', and 'inc' (in the case of censoring)}
}
\value{
An object of class 'SEQOpts'
}
\description{
Parameter Builder for SEQuential Model and Estimates
}
