% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEQuential-package.R
\docType{package}
\name{SEQuential-package}
\alias{SEQuential}
\alias{SEQuential-package}
\title{SEQuential: Target Trial Emulation}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
